requirements.txt
tweepy==3.5.0
python-dotenv==0.6.1
peewee==2.8.5
psycopg2==2.6.2